apps/api/pyproject.toml
[project]
name = "assetbridge-api"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.30",
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "pydantic-settings>=2.3",
    "litellm>=1.40",
    "yfinance>=0.2",
    "newsapi-python>=0.2",
    "pycoingecko>=3.1",
    "apscheduler>=3.10",
    "matplotlib>=3.9",
    "numpy>=1.26",
    "python-dotenv>=1.0",
    "httpx>=0.27",
    "cryptography>=42.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
